doc: Add presentation outlining the research goal
Add a very short LaTeX presentation which outlines the basic goal of the
research in a 'doc' (documentation) folder of the repository.
Furthermore, introduce a .gitignore in the doc folder to disregard LaTeX
remnants, add a literature.bib file for tracking literature references
and add the necessary resources for the presentation in 'doc/res'.