learners.py: Lossy frequency boundaries
The implementation did not enforce that several conditions are met at
the same time. Hence, adapt the comment to say so. Additionally add a
note to the config.yml file.

Finally, adapt the sampling methods in a way which ensures that the last
condition is always enforced even if sampling is selected. This comes at
the cost of potentially increasing the sampling size slightly.