learners.py: Add TF-IDF and Recall@N for said
Implement a content-based filtering approach, i.e. TF-IDF in this case
and calculate the Recall@N for said algorithm.
Hereby remove donations with no associated projects. Furthermore, add
new parameters to the config.yml including the number of Top-N non-rated
items and an options dictionary for SciKit-Learn's TfidfVectorizer.

The Recall@N is calculated by assimilating the profile of a user using
every transaction except for one and calculating the position of the
left-out transaction in a recommendation based on this profile.