*.py and config.py: Accuracy methods handling
Treat the accuracy methods like a list of accuracy methods to be
calculated instead of a dictionary of methods to be applied to the data.
This allows a dynamic selection of accuracy methods and enables a
reduction in required resources as only those calculations which are
necessary are performed.

Allow the setting of said accuracy methods via the config.yml and
provide a sensible default list.
Adapt the learners.py script accordingly.