recsys.py ContentFilters: Introduce GensimFastText
Add a method for Gensim's FastText implementation to the content-based
filtering algorithm class. Wrap the implementation in a subclass and
adjust the parent accordingly - mainly just renames and an altered
indentation level.

Make some slight adjustments to learners.py as to remove useless special
delimiter keywords.

Furthermore, adapt config.yml to include an entry for the newly added
algorithm.