meta-learners.py: Decision tree as meta-learner
Introduce a simple decision tree for predicting the position in the
Top-N accuracy set for each algorithm.

Adapt the config-meta-learners.yml to provide the necessary information
about the column names, i.e. the names of the algorithms and the name of
the accuracy methods used.

As additional small note, use itertools for nested loops and simplify a
previous invocation of one.