meta-learners.py: Simple keras model
Introduce a simplistic feed-forward model as additional meta-learner. As
fitting and predicting is different for Keras than for SKlearn, wrap the
model into a helper class to allow for it to be invoked the same way as
SKLearn's models.

In the process of introducing a neural network it has become necessary
to fill NaN values as Keras is allergic to them.

Amend the config-meta-learners.yml to accommodate the desired options
for the model.