meta-learners.py: Introduce weights for sample
Weight each sample by the difference of the two best performing
algorithm on a logarithmic scale. This penalizes false classifications
for samples for which only one algorithm provides a good estimate.

Disable this functionality for now in the config-meta-learners.yml file
and only ever use it for algorithm selection methods using
classification.