doc: Add presentation about the research project
Introduce a presentation after the fourth research progress presentation
which summarizes the research project as a whole.

Parts of the presentation are inspired by previous presentations
outlining the research project and the research progress presentations.

Amend the .gitignore file in the doc/ folder to disregard LaTeX files in
the doc/res folder as the presentation sources a table from this
location which can easily be recreated from the code mentioned in the
presentation itself.