learners.py: Revamp groupby handling
Convert the `groupby` option to a sub-options of algorithms in
`algorithms_args`. This allows to more easily enforce the names of
algorithms to be unique due to them being keys in one big dictionary.

Adapt the config.yml accordingly and remove some poor algorithms using
groupby. Document the structure of the arguments in the YAML file
itself.

Furthermore, fix the docstring of the ContentFilers class in recsys.py
as to comply with the documentation in config.yml.