meta-learners.py: Introduce user-clusters
Introduce k-Means as clustering algorithm which clusters users and
selects the best algorithm for each such cluster via a simple majority
voting system.

To simplify the implementation, switch to passing actual numpy arrays to
the meta-algorithms instead of pandas' series or dataframes. This allows
easier indexing in the context of meta-learns and avoids switching
between using index-location and simple location based lookups.